Rename file karma.config.shared.js -> karma.shared.js
