+++ karma.shared.js
